Explicitly use Player's current working directory
This allows it to be used from any directory, most importantly
root, which is required to run on boot. Fingers crossed...

Interestingly, because status.sh assumes any output is good
output, the error output was treated as successful, which
hardware.py interpreted to mean the player was playing and thus
turned on its LED. Fun!